backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
mediapipe==0.10.14
opencv-python==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
scikit-image==0.24.0
firebase-admin==6.5.0
httpx==0.27.0
python-multipart==0.0.9
python-dotenv==1.0.1
